+++ docs/scripts/views/test.vue
@@ -1,5 +1,10 @@
 <template>
-  <div></div>
+  <div>
+    <ui-form-field>
+      <label>Input:</label>
+      <ui-textfield></ui-textfield>
+    </ui-form-field>
+  </div>
 </template>
 
 <script>