+++ src/scripts/components/form-controls/form.vue
@@ -39,21 +39,26 @@ export default {
     labelRightAligned: {
       type: Boolean,
       default: false
-    }
+    },
+    // form items
+    itemMarginBottom: {
+      type: [String, Number],
+      default: 0
+    },
     // horizontal form
-    // labelWidth: {
-    //   type: Number,
-    //   default: 0
-    // },
-    // labelMarginRight: {
-    //   type: Number,
-    //   default: 0
-    // },
-    // // vertical form
-    // labelMarginBottom: {
-    //   type: Number,
-    //   default: 0
-    // }
+    labelWidth: {
+      type: [String, Number],
+      default: 0
+    },
+    labelMarginRight: {
+      type: [String, Number],
+      default: 0
+    },
+    // vertical form
+    labelMarginBottom: {
+      type: [String, Number],
+      default: 0
+    }
   },
   data() {
     return {
@@ -74,35 +79,19 @@ export default {
         'mdc-form--label-right-aligned': this.labelRightAligned
       };
     }
-    // style() {
-    //   let result = {};
-
-    //   if (this.isVertical) {
-    //     if (this.labelWidth || this.labelMarginRight) {
-    //       console.warn(
-    //         '`labelWidth`/`labelMarginRight` just for `horizontal` type form'
-    //       );
-    //     }
-
-    //     if (this.labelMarginBottom) {
-    //       result['margin-bottom'] = `${this.labelMarginBottom}px`;
-    //     }
-    //   } else {
-    //     if (this.labelMarginBottom) {
-    //       console.warn('`labelMarginBottom` just for `vertical` type form');
-    //     }
-
-    //     if (this.labelWidth) {
-    //       result['flex-basis'] = `${this.labelWidth}px`;
-    //     }
-
-    //     if (this.labelMarginRight) {
-    //       result['margin-right'] = `${this.labelMarginRight}px`;
-    //     }
-    //   }
-
-    //   return result;
-    // }
+  },
+  created() {
+    if (this.isVertical) {
+      if (this.labelWidth || this.labelMarginRight) {
+        throw new Error(
+          '`labelWidth`/`labelMarginRight` just for `horizontal` type form'
+        );
+      }
+    } else {
+      if (this.labelMarginBottom) {
+        throw new Error('`labelMarginBottom` just for `vertical` type form');
+      }
+    }
   }
 };
 </script>